refactor(modules-file): change the import/exports for type interfaces
