+++ the-toad-tribune/src/layout/Navgation.tsx
@@ -1,7 +1,7 @@
 import React, { KeyboardEvent, useState } from "react";
 import styled from "styled-components";
-import { getNewsEverything } from "../api/newsApi";
-import { NewsEverythingRequest } from "../api";
+// import { getNewsEverything } from "../api/newsApi";
+import { getNewsEverything, NewsEverythingRequest } from "../api";
 
 interface INavProps {
   renderMainLayoutPage: Function;
