+++ the-toad-tribune/src/layout/SearchResults.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import { INewsResponse } from "../api/newsApi";
-import { getNewsEverything } from "../api/newsApi";
-import { NewsEverythingRequest } from "../api";
+import { getNewsEverything, NewsEverythingRequest } from "../api";
+import type { INewsResponse } from "../api";
+// import { getNewsEverything } from "../api/newsApi";
 import PrevIcon from "../commons/prev.png";
 import NextIcon from "../commons/next.png";
