+++ the-toad-tribune/src/layout/Stonks.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { NewsProps } from "../api/newsApi";
+import type { NewsProps } from "../api";
 import { usePagination } from "../hooks";
 import {
   ArticleContentContainer,
